requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
autopep8==1.7.0
camelcase==0.2
click==8.1.3
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.5
restx-monkey==0.4.0
six==1.16.0
SQLAlchemy==1.4.42
sqlalchemy-mixins==1.5.3
toml==0.10.2
Werkzeug==2.2.2
